# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Care Platform API",
        "description": "API documentation for Pet Care Platform - connecting pet owners with reliable sitters",
        "contact": {
            "email": "admin@petcare.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sitter-profiles": {
            "get": {
                "tags": [
                    "Sitter Profiles"
                ],
                "summary": "Get list of sitter profiles (public)",
                "description": "Returns list of sitter profiles with optional filtering - no authentication required",
                "operationId": "getSitterProfilesList",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_rate",
                        "in": "query",
                        "description": "Minimum hourly rate",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_rate",
                        "in": "query",
                        "description": "Maximum hourly rate",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sitter profiles retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SitterProfile"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "404": {
                        "description": "Not Found - The requested resource was not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sitter Profiles"
                ],
                "summary": "Create new sitter profile",
                "description": "Create a new sitter profile",
                "operationId": "storeSitterProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SitterProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sitter profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sitter profile created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SitterProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid authentication token"
                    },
                    "403": {
                        "description": "Forbidden - You do not have access to this resource"
                    },
                    "404": {
                        "description": "Not Found - The requested resource was not found"
                    },
                    "409": {
                        "description": "Conflict - Sitter profile already exists for this user"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/{user_id}/sitter-profile": {
            "get": {
                "tags": [
                    "Sitter Profiles"
                ],
                "summary": "Get sitter profile by user ID (public)",
                "description": "Returns sitter profile for a specific user - no authentication required",
                "operationId": "getSitterProfileByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sitter profile retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SitterProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "403": {
                        "description": "Forbidden - User is not a sitter"
                    },
                    "404": {
                        "description": "Not Found - User or sitter profile not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SitterProfile": {
                "title": "Sitter Profile",
                "description": "Sitter Profile model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced pet sitter with 5 years of experience. Love all animals!"
                    },
                    "default_hourly_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 52.3676
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 4.9041
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "SitterProfileRequest": {
                "title": "Sitter Profile Request",
                "description": "Sitter Profile request model",
                "required": [
                    "user_id",
                    "default_hourly_rate"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced pet sitter with 5 years of experience. Love all animals!"
                    },
                    "default_hourly_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 52.3676
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 4.9041
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}